feat(guards): implements roles to current users

+++ src/modules/restaurants/restaurants.controller.ts
@@ -11,6 +11,7 @@ import {
   UseInterceptors,
   UploadedFiles,
   UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { RestaurantsService } from './restaurants.service';
@@ -21,6 +22,8 @@ import { FilesInterceptor } from '@nestjs/platform-express';
 import { AuthGuard } from '@nestjs/passport';
 import { CurrentUser } from '../auth/decorators/current-user.decorators';
 import { User } from '../auth/schemas/user.schema';
+import { RolesGuard } from '../auth/guards/roles.guard';
+import { Roles } from '../auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -32,7 +35,8 @@ export class RestaurantsController {
   }
 
   @Post()
-  @UseGuards(AuthGuard())
+  @UseGuards(AuthGuard(), RolesGuard)
+  @Roles('admin', 'user')
   async createRestaurant(
     @Body() restaurant: CreateRestaurantDto,
     @CurrentUser() user: User,
@@ -46,19 +50,32 @@ export class RestaurantsController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard())
   async findByIdAndUpdate(
     @Param('id') id: string,
     @Body() updateRestaurantDto: UpdateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    await this.restaurantsService.findById(id);
+    const restaurant = await this.restaurantsService.findById(id);
+
+    if (restaurant.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You cannot update this resturant');
+    }
+
     return this.restaurantsService.findByIdAndUpdate(id, updateRestaurantDto);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard())
   async findByIdAndDelete(
     @Param('id') id: string,
+    @CurrentUser() user: User,
   ): Promise<{ deleted: boolean }> {
     const restaurantExists = await this.restaurantsService.findById(id);
+
+    if (restaurantExists.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You cannot delete this resturant');
+    }
 
     const isDeleted = await this.restaurantsService.deleteImages(
       restaurantExists.images,
@@ -77,6 +94,7 @@ export class RestaurantsController {
   }
 
   @Put('upload/:id')
+  @UseGuards(AuthGuard())
   @UseInterceptors(FilesInterceptor('files'))
   async uploadFiles(
     @Param('id') id: string,
